fix: create note inputs with a single element
VM already adds the padding to the inputs, so creating a note with `vec![X, ZERO, ZERO, ZERO]` will actually create it as if there were 4 different inputs instead of a single one. This causes an error when trying to consume the note as it asserts that there is a single input.