chore: consistent use of EMPTY_WORD
The rust code uses ZERO as a constant for the 0 felt element.
The masm code uses ZERO for a word of four zeros, [0, 0, 0, 0].

This discrepancy made reading comments on both pieces of code confusing.
To eleminate the ambiguity this changes the references of ZERO to
EMPTY_WORD on both, and replaces uses of ZERO that refer to a single
element by the digit 0.