bugfix: allow account code to change
Changes to the account code were forbidden in the Rust code, because the
account delta nonce validation assumed the nonce would only change with
changes to the vault or storage.

Currently the account delta doesn't track changes to the account code,
so as an intermediary solution this changes the check to allow code
changes anytime. The correct solution would be to introduce the code
change to the account delta too, and introduce that as another case in
which the delta must change, and then forbid changes otherwise.