build(lib): do not run build script based on changes to `kernel_v0.rs`
https://github.com/0xPolygonMiden/miden-base/pull/887 added a build
script that builds `kernel_v0.rs`, but the script runs on changes to the
file itself. I suspect this was the reason for test times doubling
after that PR as pointed out in https://github.com/0xPolygonMiden/miden-base/issues/903.

This PR simply removes this rerun check. I doubt this check is the correct
behaviour either way, since if it reruns based on changes to the
`kernel_v0.rs`, it means it reruns every time the file is generated - which means
the build script is triggered between each invocation of `cargo nextest`
as well.